// frontend/src/app/shared/interfaces/posts/survey/survey.ts
import { Result } from "../result";

export interface Survey {
  id: string;

  linkedCardId: string;

  content: string;

  targetGroups: string;

  publicationDate: Date;

  endDate: Date;

  author: string;

  type: string;

  views: number;

  likes: number;

  answers: number;

  picture: string;

  results: Result[];

  contentSurvey: string;

  maxQuestions: number;

  questionNumber: number;
}